+++ src/1D/spline_utils_1D.jl
@@ -91,3 +91,15 @@ function spline_smoothing(lambda::Number, Delta::Number, y::Vector)
 
     return inv(diagm(ones(n)) + lambda * K) * y
 end
+
+function parse_txt_1D(path)
+    file = open(path)
+    lines = readlines(file)
+    close(file)
+
+    num_elements = parse(Int64, lines[2])
+    x = [parse(Float64, val) for val in lines[4:(3 + num_elements)]]
+    y = [parse(Float64, val) for val in lines[(5 + num_elements):end]]
+
+    return x, y
+end